+++ packages/intl-segmenter/polyfill-force.ts
@@ -1,0 +1,8 @@
+import {Segmenter} from './src/segmenter'
+
+Object.defineProperty(Intl, 'Segmenter', {
+  value: Segmenter,
+  enumerable: false,
+  writable: true,
+  configurable: true,
+})