+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.1](https://github.com/formatjs/formatjs/compare/@formatjs/cli@6.0.0...@formatjs/cli@6.0.1) (2023-02-07)
+
+**Note:** Version bump only for package @formatjs/cli
+
+# [6.0.0](https://github.com/formatjs/formatjs/compare/@formatjs/cli@5.1.13...@formatjs/cli@6.0.0) (2023-01-31)
+
+**Note:** Version bump only for package @formatjs/cli
+
+## [5.1.13](https://github.com/formatjs/formatjs/compare/@formatjs/cli@5.1.12...@formatjs/cli@5.1.13) (2023-01-30)
+
+**Note:** Version bump only for package @formatjs/cli
+
 ## [5.1.12](https://github.com/formatjs/formatjs/compare/@formatjs/cli@5.1.11...@formatjs/cli@5.1.12) (2022-12-20)
 
 **Note:** Version bump only for package @formatjs/cli