+++ packages/swc-plugin/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.19](https://github.com/formatjs/formatjs/compare/@formatjs/swc-plugin@1.4.18...@formatjs/swc-plugin@1.4.19) (2023-01-30)
+
+**Note:** Version bump only for package @formatjs/swc-plugin
+
 ## [1.4.18](https://github.com/formatjs/formatjs/compare/@formatjs/swc-plugin@1.4.17...@formatjs/swc-plugin@1.4.18) (2022-12-02)
 
 **Note:** Version bump only for package @formatjs/swc-plugin